requirements.txt
absl-py==1.4.0
altair==5.0.1
anyio==3.7.1
appdirs==1.4.4
astunparse==1.6.3
attrs==23.1.0
audioread==3.0.0
auth0-python==4.3.0
Authlib==1.2.1
backports.zoneinfo==0.2.1
blinker==1.6.2
blis==0.7.10
cachetools==5.3.1
catalogue==2.0.8
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.2.0
chromaprint==0.5
click==8.1.6
colorama==0.4.6
confection==0.1.0
contourpy==1.1.0
cryptography==41.0.2
cycler==0.11.0
cymem==2.0.7
decorator==5.1.1
exceptiongroup==1.1.2
fastai==2.7.12
fastcore==1.5.29
fastdownload==0.0.7
fastprogress==1.0.3
ffmpeg-downloader==0.2.0
ffmpeg-python==0.2.0
ffprobe==0.5
filelock==3.12.2
Flask==2.3.2
Flask-Login==0.6.2
Flask-SQLAlchemy==3.0.5
flatbuffers==1.12
fonttools==4.41.0
future==0.18.3
gast==0.4.0
gitdb==4.0.10
GitPython==3.1.32
google-auth==2.22.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
greenlet==2.0.2
grpcio==1.56.0
h11==0.12.0
h2==4.1.0
h5py==3.9.0
hpack==4.0.0
httpcore==0.13.7
httpx==0.19.0
hyperframe==6.0.1
idna==3.4
importlib-metadata==6.8.0
importlib-resources==6.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.3.1
jsonschema==4.18.4
jsonschema-specifications==2023.7.1
keras==2.9.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.4
langcodes==3.3.0
lazy_loader==0.3
libclang==16.0.0
librosa==0.8.1
llvmlite==0.38.1
Markdown==3.4.3
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.7.2
mdurl==0.1.2
mpmath==1.3.0
msgpack==1.0.5
murmurhash==1.0.9
networkx==3.1
norbert==0.2.1
numba==0.55.2
numpy==1.22.4
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==23.1
pandas==1.5.3
pathy==0.10.2
Pillow==9.5.0
pkgutil_resolve_name==1.3.10
pooch==1.6.0
preshed==3.0.8
protobuf==4.23.4
pyacoustid==1.2.2
pyarrow==12.0.1
pyasn1==0.5.0
pyasn1-modules==0.3.0
pyAudioAnalysis==0.3.14
pycparser==2.21
pydantic==1.10.11
pydeck==0.8.0
PyDejavu==0.1.3
pydub==0.25.1
pyfingerprint==1.5
Pygments==2.15.1
PyJWT==2.7.0
Pympler==1.0.1
pyparsing==3.0.9
pyserial==3.5
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
referencing==0.30.0
requests==2.31.0
requests-oauthlib==1.3.1
resampy==0.4.2
rfc3986==1.5.0
rich==13.4.2
rpds-py==0.9.2
rsa==4.9
scikit-learn==1.3.0
scipy==1.10.1
six==1.16.0
smart-open==6.3.0
smmap==5.0.0
sniffio==1.3.0
soundfile==0.12.1
soxr==0.3.5
spacy==3.6.0
spacy-legacy==3.0.12
spacy-loggers==1.0.4
spleeter==2.4.0
SQLAlchemy==2.0.18
srsly==2.4.7
streamlit==1.25.0
sympy==1.12
tenacity==8.2.2
tensorboard==2.9.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.9.3
tensorflow-estimator==2.9.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.3.0
thinc==8.1.10
threadpoolctl==3.2.0
toml==0.10.2
toolz==0.12.0
torch==2.0.1
torchvision==0.15.2
tornado==6.3.2
tqdm==4.65.0
typer==0.3.2
typing_extensions==4.7.1
tzdata==2023.3
tzlocal==4.3.1
urllib3==1.26.16
validators==0.20.0
wasabi==1.1.2
watchdog==3.0.0
Werkzeug==2.3.6
wrapt==1.15.0
zipp==3.16.1
